// src/shared/exceptions/error-codes.enum.ts
export enum ErrorCode {
  // Auth (1xxx)
  INVALID_CREDENTIALS = 'AUTH_1001',
  TOKEN_EXPIRED = 'AUTH_1002',
  TOKEN_INVALID = 'AUTH_1003',
  TOKEN_MISSING = 'AUTH_1004',
  REFRESH_TOKEN_INVALID = 'AUTH_1005',
  PASSWORD_MISMATCH = 'AUTH_1006',
  PASSWORD_SAME_AS_CURRENT = 'AUTH_1007',

  // User (2xxx)
  USER_NOT_FOUND = 'USER_2001',
  USER_ALREADY_EXISTS = 'USER_2002',
  EMAIL_ALREADY_IN_USE = 'USER_2003',
  USER_INACTIVE = 'USER_2004',
  INVALID_RECOVERY_CODE = 'USER_2005',
  RECOVERY_CODE_EXPIRED = 'USER_2006',

  // Permission (3xxx)
  ACCESS_DENIED = 'PERM_3001',
  INSUFFICIENT_PERMISSIONS = 'PERM_3002',
  FEATURE_DISABLED = 'PERM_3003',
  ROLE_NOT_ALLOWED = 'PERM_3004',

  // Validation (4xxx)
  VALIDATION_ERROR = 'VAL_4001',
  INVALID_INPUT = 'VAL_4002',
  INVALID_FILE = 'VAL_4003',
  FILE_REQUIRED = 'VAL_4004',
  INVALID_DATE_RANGE = 'VAL_4005',
  INVALID_FORMAT = 'VAL_4006',

  // Resource (5xxx)
  RESOURCE_NOT_FOUND = 'RES_5001',
  RESOURCE_CONFLICT = 'RES_5002',
  RESOURCE_ALREADY_EXISTS = 'RES_5003',

  // Club (6xxx)
  CLUB_NOT_FOUND = 'CLUB_6001',
  CLUB_ALREADY_EXISTS = 'CLUB_6002',
  CLUB_NUMBER_IN_USE = 'CLUB_6003',
  CLUB_ACCESS_DENIED = 'CLUB_6004',

  // Children (7xxx)
  CHILD_NOT_FOUND = 'CHILD_7001',
  CHILD_ACCESS_DENIED = 'CHILD_7002',

  // Profile (8xxx)
  PROFILE_NOT_FOUND = 'PROFILE_8001',
  PROFILE_ALREADY_EXISTS = 'PROFILE_8002',
  COORDINATOR_NOT_FOUND = 'PROFILE_8003',
  TEACHER_NOT_FOUND = 'PROFILE_8004',
  PROFILE_INVALID_OPERATION = 'PROFILE_8005',

  // Contact (85xx)
  CONTACT_NOT_FOUND = 'CONTACT_8501',

  // Content (86xx)
  DOCUMENT_NOT_FOUND = 'CONTENT_8601',
  MEDITATION_NOT_FOUND = 'CONTENT_8602',
  INFORMATIVE_NOT_FOUND = 'CONTENT_8603',
  EVENT_NOT_FOUND = 'CONTENT_8604',
  VIDEO_NOT_FOUND = 'CONTENT_8605',
  IMAGE_NOT_FOUND = 'CONTENT_8606',
  IDEA_NOT_FOUND = 'CONTENT_8607',
  ROUTE_NOT_FOUND = 'CONTENT_8608',
  COMMENT_NOT_FOUND = 'CONTENT_8609',
  FEEDBACK_NOT_FOUND = 'CONTENT_8610',
  PAGELA_NOT_FOUND = 'CONTENT_8611',
  WEEK_MATERIAL_NOT_FOUND = 'CONTENT_8612',

  // Pages (87xx)
  PAGE_NOT_FOUND = 'PAGE_8701',
  PAGE_CREATE_ERROR = 'PAGE_8702',
  PAGE_UPDATE_ERROR = 'PAGE_8703',
  PAGE_DELETE_ERROR = 'PAGE_8704',
  SECTION_NOT_FOUND = 'PAGE_8705',
  SECTION_CREATE_ERROR = 'PAGE_8706',
  SECTION_UPDATE_ERROR = 'PAGE_8707',
  SECTION_DELETE_ERROR = 'PAGE_8708',

  // Media (88xx)
  MEDIA_NOT_FOUND = 'MEDIA_8801',
  MEDIA_UPLOAD_ERROR = 'MEDIA_8802',
  MEDIA_DELETE_ERROR = 'MEDIA_8803',
  MEDIA_FIELD_MISSING = 'MEDIA_8804',
  MEDIA_FILE_NOT_FOUND = 'MEDIA_8805',

  // Internal (9xxx)
  INTERNAL_ERROR = 'INT_9001',
  DATABASE_ERROR = 'INT_9002',
  EXTERNAL_SERVICE_ERROR = 'INT_9003',
  FILE_UPLOAD_ERROR = 'INT_9004',
  EMAIL_SEND_ERROR = 'INT_9005',
  S3_DELETE_ERROR = 'INT_9006',
  S3_UPLOAD_ERROR = 'INT_9007',
  TRANSACTION_ERROR = 'INT_9008',
}